requirements.txt
discord.py==2.0.1
Unidecode==1.1.2
